+++ example/test_bot.dart
@@ -7,6 +7,18 @@ void main() async {
   /// Creates the bot instance
   final bot = Bot(
     Platform.environment["BOT_TOKEN"]!,
+    loggerOptions: LoggerOptions(
+      request: true,
+      requestHeader: true,
+      requestBody: true,
+      responseHeader: true,
+      responseBody: true,
+      error: true,
+      methods: [
+        APIMethod.sendMessage,
+        APIMethod.getMe,
+      ],
+    ),
   );
 
   bot.command('hi', (ctx) async {
@@ -15,6 +27,11 @@ void main() async {
   });
 
   bot.start();
+
+  bot.command('me', (ctx) async {
+    final user = await ctx.api.getMe();
+    await ctx.reply('My name is ${user.firstName}');
+  });
 }
 
 void handleName(MessageContext ctx) async {